requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
anthropic>=0.20.0
numpy>=1.24.0
python-dotenv>=1.0.0